+++ lib/index.ts
@@ -13,6 +13,31 @@ import { JSONResume } from './schema';
 const ajv = new Ajv();
 const validator = ajv.compile(schema);
 
-export function isValid(data: object): data is JSONResume {
+/**
+ * Checks whether the given object is a valid JSON Resume.
+ *
+ * This function does not return any information if validation fails (for that,
+ * use {@link validate}). This function is mostly useful as a type guard when
+ * such additional information is unnecessary.
+ *
+ * @param data the object to check
+ */
+export function isValid(data: any): data is JSONResume {
   return validator(data) as boolean;
+}
+
+/**
+ * Checks whether the given object is a valid JSON Resume, returning validation
+ * errors on failure.
+ *
+ * @param data the object to check
+ * @returns any errors encountered during validation, or an empty array if valid
+ */
+export function validate(data: any): string[] {
+  if (!validator(data) && validator.errors) {
+    return validator.errors.map(
+      e => `${e.dataPath}: ${e.message ? e.message : e.keyword}`,
+    );
+  }
+  return [];
 }